Add files via upload
All data is accessed from mlb.com via Selenium.  For each pitcher and batter selected the following stats for the batter are provided: season, left landed (for season), right handed (for season), against selected pitcher for season and career.  A print screen is included.  Tkinter was used for buttons and to display data.  This program is useful while watching a game to determine how the batter has performed against a pitcher.  Performance needs to be improved in future releases.  Also for future releases,  the addition of postseason stats for the batter plus relevant stats for the pitcher.