Add files via upload
This is an example of the "The_Pennant_Race.ipynb".  Lists standings for he first and second place team on a given day. Also, what win/loss record is needed to catch the first place team given the first place teams record for the rest of the season.  Plus it lists over 500 opponents for the rest of the season and how many are remaining.  Answers the questions, What will it take to catch the first team and who has the easier schedule?